libs/shared-ocr/pyproject.toml
[project]
name = "shared-ocr"
version = "0.1.0"
description = "OCR service for reading insurance card information using Gemini 2.5 Flash"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.8.3",
    "pillow>=10.4.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["libs/shared_ocr"]